requirements.txt
boto==2.38.0
celery==3.1.18
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.8.1
django-allauth==0.19.1
django-extensions==1.5.2
djangorestframework==3.1.1
gunicorn==19.3.0
oauthlib==0.7.2
opbeat==2.1
psycopg2==2.6
python3-openid==3.0.5
pytz==2015.2
requests==2.6.1
requests-oauthlib==0.4.2
six==1.9.0
tzlocal==1.1.3
whitenoise==1.0.6
youtube-dl==2015.4.17
